[xla:gpu] Extract CommandExecutor in a separate target
