+++ xla/backends/gpu/runtime/BUILD
@@ -94,6 +94,41 @@ cc_library(
 )
 
 cc_library(
+    name = "command_executor",
+    srcs = ["command_executor.cc"],
+    hdrs = ["command_executor.h"],
+    deps = [
+        ":annotation",
+        ":command",
+        ":command_state",
+        ":thunk",
+        "//xla:util",
+        "//xla/runtime:buffer_use",
+        "//xla/runtime:execution_graph",
+        "//xla/runtime:resource_use",
+        "//xla/service:buffer_assignment",
+        "//xla/stream_executor:command_buffer",
+        "//xla/tsl/platform:env",
+        "//xla/tsl/platform:errors",
+        "//xla/tsl/platform:statusor",
+        "@com_google_absl//absl/algorithm:container",
+        "@com_google_absl//absl/base:core_headers",
+        "@com_google_absl//absl/container:btree",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/container:flat_hash_set",
+        "@com_google_absl//absl/container:inlined_vector",
+        "@com_google_absl//absl/log",
+        "@com_google_absl//absl/log:check",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/strings:str_format",
+        "@com_google_absl//absl/types:span",
+        "@tsl//tsl/profiler/lib:scoped_annotation",
+    ],
+)
+
+cc_library(
     name = "command_buffer_cmd",
     srcs = ["command_buffer_cmd.cc"],
     hdrs = ["command_buffer_cmd.h"],
@@ -107,6 +142,7 @@ cc_library(
         ":collective_permute_thunk",
         ":collective_thunk",
         ":command",
+        ":command_executor",
         ":command_state",
         ":copy_thunk",
         ":custom_call_thunk",
@@ -191,6 +227,7 @@ xla_test(
     deps = [
         ":command",
         ":command_buffer_cmd",
+        ":command_executor",
         ":command_state",
         ":copy_thunk",
         ":shaped_slice",
@@ -240,6 +277,7 @@ cc_library(
         ":collective_thunk",
         ":command",
         ":command_buffer_cmd",
+        ":command_executor",
         ":command_state",
         ":conditional_thunk",
         ":copy_thunk",
@@ -375,7 +413,7 @@ cc_library(
     hdrs = ["command_buffer_thunk.h"],
     deps = [
         ":command",
-        ":command_buffer_cmd",
+        ":command_executor",
         ":command_state",
         ":sequential_thunk",  # build_cleaner: keep
         ":thunk",
@@ -409,6 +447,7 @@ xla_test(
         ":command",
         ":command_buffer_cmd",
         ":command_buffer_thunk",
+        ":command_executor",
         ":dynamic_slice_thunk",
         ":gpublas_lt_matmul_thunk",
         ":memset_thunk",
@@ -461,6 +500,7 @@ xla_test(
         ":command",
         ":command_buffer_cmd",
         ":command_buffer_thunk",
+        ":command_executor",
         ":thunk",
         "//xla:types",
         "//xla:xla_data_proto_cc",
@@ -1109,9 +1149,9 @@ xla_test(
     srcs = ["kernel_thunk_test.cc"],
     backends = ["gpu"],
     deps = [
-        ":command_buffer_cmd",
         ":command_buffer_cmd_emitter",
         ":command_buffer_thunk",
+        ":command_executor",
         ":kernel_thunk",
         ":sequential_thunk",
         ":thunk",
@@ -1665,9 +1705,9 @@ xla_test(
     deps = [
         ":collective_broadcast_thunk",
         ":collective_thunk",
-        ":command_buffer_cmd",
         ":command_buffer_cmd_emitter",
         ":command_buffer_thunk",
+        ":command_executor",
         ":sequential_thunk",
         ":thunk",
         ":thunk_proto_cc",
@@ -1751,9 +1791,9 @@ xla_test(
     deps = [
         ":collective_permute_thunk",
         ":collective_thunk",
-        ":command_buffer_cmd",
         ":command_buffer_cmd_emitter",
         ":command_buffer_thunk",
+        ":command_executor",
         ":sequential_thunk",
         ":thunk",
         ":thunk_proto_cc",
@@ -2128,9 +2168,9 @@ xla_test(
     use_legacy_runtime = True,
     deps = [
         ":collective_thunk",
-        ":command_buffer_cmd",
         ":command_buffer_cmd_emitter",
         ":command_buffer_thunk",
+        ":command_executor",
         ":recv_thunk",
         ":send_thunk",
         ":sequential_thunk",
@@ -2513,8 +2553,8 @@ xla_cc_test(
     srcs = ["for_all_thunks_test.cc"],
     tags = ["gpu"],
     deps = [
-        ":command_buffer_cmd",
         ":command_buffer_thunk",
+        ":command_executor",
         ":conditional_thunk",
         ":dynamic_slice_thunk",
         ":sequential_thunk",
@@ -3455,9 +3495,9 @@ cc_library(
     srcs = ["command_buffer_conversion_pass.cc"],
     hdrs = ["command_buffer_conversion_pass.h"],
     deps = [
-        ":command_buffer_cmd",
         ":command_buffer_cmd_emitter",
         ":command_buffer_thunk",
+        ":command_executor",
         ":conditional_thunk",
         ":copy_thunk",
         ":custom_call_thunk",
