+++ xla/backends/gpu/runtime/command.cc
@@ -34,9 +34,11 @@ bool IsCollectiveCommand(CommandType type) {
     case CommandType::kAllGatherCmd:
     case CommandType::kAllReduceCmd:
     case CommandType::kAllToAllCmd:
-    case CommandType::kReduceScatterCmd:
     case CommandType::kCollectiveBroadcastCmd:
     case CommandType::kCollectivePermuteCmd:
+    case CommandType::kReduceScatterCmd:
+    case CommandType::kRecvCmd:
+    case CommandType::kSendCmd:
       return true;
     default:
       return false;
